Move sub_repo_perms_store to enterprise
