+++ enterprise/cmd/gitserver/shared/shared.go
@@ -8,6 +8,7 @@ package shared
 import (
 	"github.com/sourcegraph/log"
 	srp "github.com/sourcegraph/sourcegraph/enterprise/internal/authz/subrepoperms"
+	edb "github.com/sourcegraph/sourcegraph/enterprise/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 )
@@ -15,6 +16,7 @@ import (
 func EnterpriseInit(db database.DB) {
 	logger := log.Scoped("enterprise", "gitserver enterprise edition")
 	var err error
+	database.SubRepoPermsWith = edb.SubRepoPermsWith
 	authz.DefaultSubRepoPermsChecker, err = srp.NewSubRepoPermsClient(db.SubRepoPerms())
 	if err != nil {
 		logger.Fatal("Failed to create sub-repo client", log.Error(err))
